moved login component inside of the components folder
